Call cmd_report only on replied messages
Otherwise there is nothing to report

+++ bouncerbot/bouncerbot.py
@@ -7,8 +7,12 @@ class BouncerBot(Bot):
         super().__init__(config['bot'])
 
     def cmd_report(self, message):
-        replied = self.get_replied_message(message)
-        print("Reporting %s" % replied)
-        self.delete_message(replied.to_id.channel_id, replied.id)
+        if self.is_reply(message):
+            replied = self.get_replied_message(message)
+            print("Reporting %s" % replied)
+            self.delete_message(replied.to_id.channel_id, replied.id)
+        else:
+            # FIXME: describe that we need replied message
+            pass
 
     cmd_spam = cmd_report
